improve(cli): move newCmd functions
Most Go CLIs follow this pattern where the function which creates the
cobra.Command is in the same file as the implementation of that command.
