+++ internal/cmd/logout.go
@@ -3,8 +3,28 @@ package cmd
 import (
 	"errors"
 
+	"github.com/spf13/cobra"
+
 	"github.com/infrahq/infra/internal/logging"
 )
+
+func newLogoutCmd() *cobra.Command {
+	var purge bool
+
+	cmd := &cobra.Command{
+		Use:     "logout",
+		Short:   "Log out of Infra",
+		Example: "$ infra logout",
+		Group:   "Core commands:",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return logout(purge)
+		},
+	}
+
+	cmd.Flags().BoolVar(&purge, "purge", false, "remove Infra host from config")
+
+	return cmd
+}
 
 func logout(purge bool) error {
 	config, err := readConfig()
